services/order/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
asyncpg
pydantic
pydantic-settings==2.1.0
aio-pika==9.3.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
prometheus-fastapi-instrumentator
